# fix wrong offset for VERIFY_PATTERN_NO_HDR