hdr points the beginning of a block filled with a pattern

Signed-off-by: Oleg Krasnov <ko-85@bk.ru>